Part 1 done
